+++ haystack_experimental/tools/__init__.py
@@ -2,6 +2,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from .tool_component_descriptor import create_tool_parameters_schema
+from .component_schema import create_tool_parameters_schema
 
 __all__ = ["create_tool_parameters_schema"]
